move About widget into body/ folder, fix imports in several files
